daredevil_backend/pyproject.toml
[project]
name = "daredevil-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "authlib>=1.6.3",
    "cryptography>=45.0.7",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "logfire[asyncpg,fastapi]>=4.6.0",
    "opentelemetry-instrumentation-asyncpg>=0.57b0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "rich>=14.1.0",
    "ruff>=0.13.0",
]

[tool.pytest.ini.options]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra"
    ]
testpaths = "tests"

markers = [
    "smoke: subset of tests",
    "exception: check for expected exceptions"
]

